% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_effort_surface.R
\name{generate_effort_surface}
\alias{generate_effort_surface}
\title{Generate a Sampling Effort Proxy Surface}
\usage{
generate_effort_surface(country = "Pakistan", resolution = 10000)
}
\arguments{
\item{country}{Character. Name of the country (e.g., "Pakistan").}

\item{resolution}{Numeric. Output raster resolution in meters (default = 10000).}
}
\value{
A SpatRaster object where higher values indicate
higher probability of human sampling effort (closer to roads).
}
\description{
Downloads road data for a given country and calculates
distance to roads to create a proxy for human sampling effort.
}
